# backend/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/chat:
        post:
            tags:
                - Chat
            operationId: Chat_Chat
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/aigc.v1.ChatRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aigc.v1.ChatReply'
    /api/conversation:
        get:
            tags:
                - Conversation
            operationId: Conversation_GetConversation
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aigc.v1.GetConversationReply'
        post:
            tags:
                - Conversation
            operationId: Conversation_CreateConversation
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/aigc.v1.EmptyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aigc.v1.CreateConversationReply'
    /api/conversation/{conversationId}/context:
        get:
            tags:
                - Conversation
            operationId: Conversation_GetConversationContext
            parameters:
                - name: conversationId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aigc.v1.GetConversationContextReply'
    /api/login:
        post:
            tags:
                - User
            operationId: User_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/aigc.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aigc.v1.LoginReply'
    /api/logout:
        post:
            tags:
                - User
            operationId: User_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/aigc.v1.LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aigc.v1.LogoutReply'
    /api/register:
        post:
            tags:
                - User
            operationId: User_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/aigc.v1.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aigc.v1.RegisterReply'
    /api/sendCode:
        post:
            tags:
                - User
            operationId: User_SendCode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/aigc.v1.SendCodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aigc.v1.SendCodeReply'
    /api/userInfo:
        get:
            tags:
                - User
            operationId: User_GetUserInfo
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/aigc.v1.GetUserInfoReply'
components:
    schemas:
        aigc.v1.ChatData:
            type: object
            properties:
                conversationId:
                    type: string
                content:
                    type: string
        aigc.v1.ChatReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    $ref: '#/components/schemas/aigc.v1.ChatData'
        aigc.v1.ChatRequest:
            type: object
            properties:
                message:
                    type: string
                conversationId:
                    type: string
                filePath:
                    type: string
        aigc.v1.ContextData:
            type: object
            properties:
                role:
                    type: string
                content:
                    type: string
        aigc.v1.ConversationData:
            type: object
            properties:
                cid:
                    type: string
                calMessage:
                    type: string
        aigc.v1.CreateConversationReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                conversationId:
                    type: string
        aigc.v1.EmptyRequest:
            type: object
            properties: {}
        aigc.v1.GetConversationContextReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                context:
                    type: array
                    items:
                        $ref: '#/components/schemas/aigc.v1.ContextData'
        aigc.v1.GetConversationReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/aigc.v1.ConversationData'
        aigc.v1.GetUserInfoReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                email:
                    type: string
                message:
                    type: string
        aigc.v1.LoginReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                token:
                    type: string
                message:
                    type: string
        aigc.v1.LoginRequest:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
        aigc.v1.LogoutReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
        aigc.v1.LogoutRequest:
            type: object
            properties: {}
        aigc.v1.RegisterReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    type: string
        aigc.v1.RegisterRequest:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
                code:
                    type: string
        aigc.v1.SendCodeReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                data:
                    type: string
        aigc.v1.SendCodeRequest:
            type: object
            properties:
                email:
                    type: string
tags:
    - name: Chat
    - name: Conversation
    - name: User